+++ app/src/main/java/ru/calloop/pikabu_demo/ui/repositories/SharedPreferences/ISessionPreferenceRepository.java
@@ -3,7 +3,7 @@ package ru.calloop.pikabu_demo.ui.repositories.SharedPreferences;
 public interface ISessionPreferenceRepository {
     void startUserSession(int accountId);
 
-    long getAccountId();
+    int getAccountId();
 
     boolean sessionStarted();
